Fixed regression in clone caused by refactoring
The Init() function was returning a new Client, but the repo that still
gets passed around and hasn't been completely replaced yet wasn't updated
to the Client's GitDir, causing it to crash downstream when the repo is used
in delta calculations.

This creates a new repo object immediately after Init in Clone in order to
make sure it's pointing at the correct place. (The long term solution is to
get rid of *libgit.Repository entirely.)