Refactored code into cmd and git packages
Git is for types and implementations that take arguments and return
in terms of Go object types, while cmd is for parsing command line options
and wrapping around git library.

This first commit mostly just involves moving the client to git/, the commands
to cmd/, and then making any changes necessary to get it to compile. There's
still a lot of refactoring to do in terms of moving the implementation details
of the cmd/ implementations into git/, and updating them to properly use flag
where appropriate.