Added git unpack-objects command
The work was mostly already done for fetch, but this
adds the git subcommand as a stand-alone command with
the same parameters as the official git client.

This should make it easier to debug why some fetches
go wrong, since we can save the packfile and go straight
to the unpacking step.