Only request refs/heads refspecs, not everything
We were previously requesting every single refspec, including
ones that aren't branches. This causes the packfile that's
downloaded to be larger than it needs to be. We now only say
that we want HEAD and refs/heads refspecs to make cloning
(slightly) faster.

Also advertise that we support the OFS_DELTA capability.
go.googlesource.com was sending OFS_DELTAs whether we advertised
the capability or not. Since we've done the work of implementing
it so that we can go get golang.org/x/... repos, we might
as well let anyone use it, and not just servers that are
actively doing evil.