Implemented git index-pack and refactored unpack-objects
This adds support for git-index-pack, which is the first step
towards directly reading packfiles, instead of needing to extract
them loosely in order to use them. (Once done, that should
significantly increase the speed of cloning, which currently
needs to write one file per object.) This is also the first
step towards getting rid of the third party git library, which
is only currently used for dealing with objects in packfiles
created by other git clients.

The unpack-objects code was also refactored in order to share
more code between REF_DELTA and OFS_DELTA, which was long
over-due, but also necessary in order to also share the delta related
code with git-index-pack.