Added support for packfiles in GetObject
Now that index-pack is supported, we have everything we
need to directly read objects from packfiles, instead
of having to unpack them before accessing them.

This restores interoperability with the official git
client that was lost when I started ripping out the
libgit library, but does it without depending on any
third party libraries.