Fixed bug where write-tree sometimes wrote the tree wrong.
Indexes store all files in a flat namespaces of their absolute
path, while trees recursively store other trees with paths
relative to the parent.

The code which tries to convert between the two had a bug
where it would sometimes get confused by the last entry
in a directory while parsing the index, resulting in it
sometimes duplicating files into trees that they shouldn't
be contained in.

This hopefully fixes it, although the code still isn't
very well written and I'm still not convinced is always
correct.