Optimized getting tree objects a little
TreeID.GetAllObjects was still using the libgit hack, which was
adding considerable overhead when trying to access objects in
trees.

This updates the code to directly parse the tree entry, and
also adds a caching layer to the client to keep track of which
packfiles have which objects in order to speed repeated access.