Began refactoring merge-base ancestory to not use libgit.
This starts getting rid of unnecessary usage of libgit, but unfortunately
comes with a fairly significant performance hit. The slice based implementation
of Ancestors is significantly worse than the linked-list based representation
that the old git library used.

This will need further improvements to be useable before it can be merged into
master.