Refactor git status to be written in terms of plumbing commands
The previous dgit status command was implemented before git ls-files and
git diff-index were implemented. This reimplements it in terms of the
plumbing, so that it's more reliable (and also seems to be faster now.)

The status command is also now split up into cmd/git rather than being
hardcoded in cmd/, making it more useable as a library.

This required implementing a couple more git ls-files options, which now
supports --unmerged and --directory.