Use a callback to walk the history for "git log"
The previous implementation required getting all commits, and then
iterating over them. This uses a callback to print them as it's
walking the tree.

This is faster, particularly on Plan 9, where there was a pause
getting the entire history before beginning to print the first commits.

It may result in the log being shown in a different order than the
official git client for some histories.