Avoid hashing while doing git status
git status was hashing every file on the filesystem to determine its
status. The updates checks to only be based on the stat information.

Plan9 only provides mtime with 1 second precision, so on Plan 9 we pretend
that the Qid.Version is the number of nanoseconds in order to ensure that
changes don't get missed.