Fix bug where status and diff show unmodified files
If a file is modified according to stat information, they still need
to be hashed in order to catch the case where they were modified, and
then modified back to the original so the stat information is off.

The stat info does, however, allow us to short-circuit out of files that
are definitely not modified.