Do not panic when .gitconfig doesn't exist.
GetAuthor and GetCommitter paniced if opening .gitconfig returned an
error, this resulted in "git clone" (and git init, and various other
commands) failing when trying to maintain the reflog. It now returns
(and handles) an error instead, and git commit prints a warning and
instructions if the config file doesn't exist.

Fixes #9