Make output go to stdout if there are no arguments
The real git client prints output information to stdout if there
are no arguments passed on the command line, not stderr. The test
suite from git causes spurious failures when testing whether the
test suite is working because of tihs.

Change default to go to stdout, not stderr to match the official git
client so we can run their test suite.