Improve git config command (#40)
The git config command now recognizes more switches. The type canonicalization switches (--type, --bool, etc.) are recognized but ignored avoiding errors when "--bool" or other switches are not present in the configuration. Set is implied by the present of a key and a value. There is a new "--unset" functionality to unset a config value.