Fix ls-files so that it doesn't show sibling files (#62)
The official git test suite was catching a case where when you do a "git ls-files foo" and there is a sibling file called "foobar". It was listing "foobar" even though it is not a child of foo. Resolving this fixes one half of the problem with the t0000-basic.sh 'update-index D/F conflict' test case.