Add inode to stat information which is compared
This allows use to include a hash of the file when comparing
with diff-files.

Also fixed a bug where HashFile wasn't working correctly with
symlinks (which it turns out was the real cause of the problem causing
some basic tests to fail, not the fact that the file isn't supposed to
be hashed).