Add special case for merge-base (#104)
Add special case for a merge-base containing only
two commits, where if one commit is an ancestor of
the other that is taken as the merge-base without
doing a search. This makes `dgit merge-base A B`
and `dgit merge-base B A` return the same commit in
these cases (which is what real git does), rather
than returning the zero commit in one but the right
commit in the other.