Added ability to push new branches
The .git/config must be manually set up for the branch ("--set-upstream-origin" is not
implemented), but the code can now handle pushing to non-existent branches in "git push".
It will use the known branch heads to determine what to put in the pack file, and send
a diff based on what it can determine from the closest branches that it knows.