Add support for git ls-remote command and preliminary support for protoco v2
This adds the git ls-remote command, and does it in a way that should be easier
to re-use with other non-http protocols eventually. It also adds support for the
git v2 protocol (but only in ls-remote) and eventually expand for proper fetch-pack
support, rather than the SmartHttpReceiver hack that currently exists.

The existing code is mostly untouched in order to not break things that are
currently working until the re-factoring is complete.