Add support for git ls-remote command and preliminary support for protocol v2 (#148)
This adds the git ls-remote command, and does it in a way that should be easier
to re-use with other non-http protocols than the SmartHttpReceiver which currently
exists. It also adds support for the git v2 protocol (but currently only in ls-remote.)

ls-remote supports the git protocol, smart http protocol, and ssh protocol in both the
v1 and v2 variants but not the dumb http protocol or local filesystem remotes.