Implement preliminary support for non-HTTP protocols (#151)
This adds the git-fetch-pack command with support for all protocols supported by git.

It adds support for git protocol versions 1 and 2 over ssh, git, http, or the local filesystem. (The latter requires a git-upload-pack implementation in your path, which isn't included with dgit.)

There is no support for incrementally downloading objects, so this in only suitable for clone, not fetch/pull yet. (Clone is also not yet updated to use the new fetch-pack implementation.)