Improve "git checkout -b" with staged files.
This improves the behaviour of "git checkout -b" when files are staged
to pass the official git test which tests for that behaviour. It makes
sure that staged changes remain staged after creating the new branch.

It also improves "diff" and "status" behaviour to more closely match
git, in that stat info is refreshed before doing the diff/status.