Refresh stat info when doing high level diff or status command
This refreshes the index stat info before doing a "diff" or
"status" command, which appears to be more in line with what
the official git client does.

This avoids diffs where the diff says a file was changed but
has an empty diff, and status claiming that unmodified files are
modified.