Add config cache to client to improve checkout time
Since the protectHFS and protectNTFS config options were added to dgit
in order to pass more of the official git tests, readtree has been
looking them up for each file.  This results in the global and local
configs each getting parsed twice per file in the repo when doing a
checkout.  (This is obviously particularly bad for large repos.)

This adds 2 levels of caching to looking up a config option:

1. The parsed config files (both local and global) are stored in the
   Client so that they only need to be parsed once per command invocation.
2. A map is used to store values that have already been looked up
   since they shouldn't change between git invocations (with the exception
   of the "git config" command.)

read-tree is currently the only thing updated to use the new
client.GetConfig function, but it would probably be a good idea to use
it consistently throughout dgit.