Fix panic when retrieving a size 0 object from a pack file. (#163)
When attempting to look up an object of size 0 (ie. an empty file)
in a git packfile, the SectionReader allocated would return EOF after
0*3 = 0 bytes are read, causing a panic when trying to read the header
for the zlib stream.

This avoids the overhead of reading the stream entirely by just directly
allocating a 0 byte slice. This should fix the panic when trying to clone
https://github.com/golang/go.