Clean up c.GetConfig changes (#172)
PR #170 changed some functions in client.go to use c.GetConfig. The problem is that GetConfig returns empty string if config does not exist.

This does not check the config presence on the system, but at least does two separate checks for empty user/email options. It'll still return NoGlobalConfig if options are empty even if the config exists.