Fix the flags with optional values in the status command (#198)
The status command has --porcelain and two other flags that can be
given on the command-line with an optional value. The git manual
provides default values in case the value is not provided. Dgit
currently requires an explicit value because of limitations in the
way that the flag package works in Go.

This works around the problem by pre-processing the set of subcommand
arguments and inserting the default values when the bare flag is
provided on the command line.

Fixes #27.