Fix the "init --bare", the detection of bare repos and a bug with ls-remote (#199)
Initializing a bare repository didn't set the configuration properly. Also, the detection of a bare repo didn't work and falsely reporting that it wasn't a git repo. Finally, ls-remote was not outputting anything with the "-q" option, which is only supposed to squelch the "From " line that gets echoed to stderr.