Add support for reflog expire --expire=now --all
This adds a hack for supporting "reflog expire --expire=now --all".

When that exact combination of parameters exist, the logs directory
is walked and all files are simply truncated.

This allows us to run the setup of the t1450-fsck test from the
official git test suite. (There's a catch-22 where fsck tests depend
on reflog and reflog tests depend on fsck, but this adds just enough
to try adding fsck.)