Do not chdir while doing an init, again (#225)
Do not chdir while doing an init, again

This was only done so that tests were run in the right directory,
but tests should be the ones doing the Chdir themselves. Changing
the directory causes paths to be evaluated relative to the wrong
directory when handling user input after an init, such as while
doing a clone.

However, the reset command invoked from of a clone needs to be able
to make paths relative.  As result, we still need to chdir within
clone, but restore it for the caller after we're done.