Add some t5501-fetch tests from the official suite (#237)
This adds the first three sanity test cases of t5501-fetch from the official test suite and fixes an off-by-1 error that was causing looking up objects that start with "0" in pack files from failing..

Note that there is a special patch against the official git-upload-pack to prevent it from invoking dgit for the pack-objects, since the subcommand is not yet ready to support what is required from upload-pack.

The run-tests script has been changed so that it no longer invokes make. The reason for this is that this form is much easier to modify to add "-d" and "-v" beside the necessary suites to get more details when tests are failing in travis. Also, you can add DGIT_TRACE in there to get the inner logs of dgit when tests fail.