Add tag tests for fsck
This gets the git tag tests working for fsck.

It includes a lot of hacks, because the git tests
regarding tags are incredibly inconsistent about whether
things go to stdout, stderr, and whether they cause a
non-zero exit status or not, but this seems to get them
all to pass, at least for now.