Fix bug where getting objects for a tree always tried to get empty Sha1 (#249)
A refactoring done in df1a505c resulted in recursively calling GetAllObjectsExcept
on a variable that was not yet initialized. This resulted in trying to
get objects on the empty Sha1, incorrectly. Fix variable reference to
the proper variable name.

This also removes the default-acl test from the official test suite, which was
only passing incidentally and is now failing with this bug fix.