Improve index pack (and hence, clone) performance (#260)
This fixes some hacks to significantly improve the performance of index-pack (cloning the dgit repo on 9front went from ~160s to ~20s). Since we've forked the compress/zlib repository anyways, the signature is changed to always require a flate.Reader instead of an io.Reader. This means that reading a zlib stream doesn't overshoot the reader, but only works on things that have a ReadByte method. Since files don't implement ReadByte, zlib also grew the ability to tell you how many compressed bytes it read, meaning files can be wrapped in a bufio and then Seeked back to the proper place.

In the refactoring the ability to simultaneously index as the pack is read from the network was lost, but this can be re-added in the future, for now the performance gains of not reading 1 byte at a time or seeking back looking for the digest more than makes up the difference.