Cache objects for delta resolution when indexing back
Delta resolution used to use a cache of objects, which was removed
because on large repos the cache was growing too large and using up
the memory. This splits indexing into 2 phases so that we can use
a cache again for large repos. The first phase hashes objects and notes
how many times they're used as the base for deltas, the second phase
caches the objects and only stores them in the cache until the last
delta that uses them is resolved.