More IndexPack improvements (#267)
This makes a variety of improvements to improve the performance of git index-pack under dgit.

    Deltas are now io.Readers so that they can be directly streamed into anything that takes an io.Reader without first having to calculate the entire delta content and store the result on the heap.
    Various places that were using make([]byte, 1) now take a flate.Reader instead of an io.Reader to read a byte. This generates much less garbage for the GC to collect. We're also more careful about passing flate.Readers around where needed instead of io.Readers
    Progress information is now sent to stderr at most once per 3 seconds.
    Indexing is now split into 2 phases. The first hashes normal commit/tree/blob/tag entries and notes how many deltas are used against them, the second resolves deltas.
    Getting objects while indexing no longer uses GetObject to resolve deltas. GetObject requires entries in the index to be sorted, and we now only sort once at the end.
    An LRU cache is used to cache objects while indexing.

My first attempt at re-adding caching cached tried to mark everything with deltas in the first phase, and then purge them after all deltas against them were resolved, but that could only get to about 80% of indexing github.com/Perl/perl5 (before getting to the delta resolution) on my 2Gb vultr node before running out of memory, so I wrote an LRU cache implementation with a cap on the amount of memory used until it starts evicting content (rather than the number of entries), but that also kept running out of memory without a very tiny cap. I assumed my code was buggy and leaking memory and switched to package github.com/hashicorp/golang-lru and.. also kept running out of memory. I think Go's garbage collector just isn't very amenable to an LRU cache. For now it's still using the hashicorp/golang-lru implementation with a very small (250 entry) cap on the number of entries (larger for small repositories), which is small enough that I can index the perl pack, although the memory consumption is larger than it should be.

The result of all this is that indexing the pack from github.com/Perl/perl5 takes ~8.5 minutes (down from ~20 minutes) on my poor little vultr node and there should be a proportional improvement in cloning time. (There's still a lot of room for improvement compared to git which takes about 1.5 minutes and doesn't bloat up the memory like this does, but this is probably a good enough improvement to commit.)

unpack-objects is also updated to use the same infrastructure as index-pack so that the older, less efficient delta resolution code can be removed.

This also fixes #262 as a side-effect of point 4.