Add support for t5300 pack-objects test (#268)
This adds support for the first few tests of t5300, which
tests pack/unpack objects in the official git client. It passes
until the point where it compares the delta flavours.

This mostly involves slightly better support for the
GIT_OBJECT_DIRECTORY environment variable, naming the packs correctly
(with the trailer as a suffix) and parsing some command line options.