Refactor push to use send-pack plumbing
This refactors "git push" to use the send-pack plumbing
rather than the hack which was previously in place to
allow pushing in the early days of dgit. The result should
be that it will work with all transport protocols that are
supported, and not just HTTP.

It also splits the push command into the same git/cmd package
distinction for parsing command line and implementing the logic,
another leftover from the fact that push was one of the earliest
things implemented in dgit.

Also fixes a bug where Linux wasn't loading SSH keys because of
incorrect build tags.