Fix regressions in git push, again
This fixes some regressions in git push caused by porting the
code to use git-send-pack instead of the old PackNegotiator https only
hack.

Our delta calculation isn't robust enough for the real world, so don't
use it in send-pack, and we weren't taking the "new branch" case into
consideration when checking if a push needs to be forced.